# pkg/api/generated/client/api/openapi.yaml
openapi: 3.0.0
info:
  description: spirits is a turn-based battle royale game
  title: spirits
  version: 0.0.1
servers:
- description: The hosted spirits gameserver
  url: https://oh-great-spirits.herokuapp.com
paths:
...
  /sessions:
    get:
      description: Watch Session
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful Session watch
      security:
      - oidc:
        - spirits::sessions.write
        - spirits::sessions.read
    post:
      description: Create a Session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
        description: Session to create
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful Session create
      security:
      - oidc:
        - spirits::sessions.write
  /sessions/{sessionName}:
    delete:
      description: Watch Session
      parameters:
      - description: Session name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Session name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful Session delete
      security:
      - oidc:
        - spirits::sessions.write
    get:
      description: Watch Session
      parameters:
      - description: Session name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Session name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful Session watch
      security:
      - oidc:
        - spirits::sessions.write
        - spirits::sessions.read
    put:
      description: Update Session
      parameters:
      - description: Session name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Session name
          minLength: 1
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
        description: Session to update
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful Session update
      security:
      - oidc:
        - spirits::sessions.write
  /sessions/{sessionName}/teams:
    get:
      description: Watch Team
      parameters:
      - description: Team name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successful Team watch
      security:
      - oidc:
        - spirits::sessions:teams.write
        - spirits::sessions:teams.read
    post:
      description: Create a Team
      parameters:
      - description: Team name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
        description: Team to create
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successful Team create
      security:
      - oidc:
        - spirits::sessions:teams.write
  /sessions/{sessionName}/teams/{teamName}:
    delete:
      description: Watch Team
      parameters:
      - description: Team name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      - description: Team name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successful Team delete
      security:
      - oidc:
        - spirits::sessions:teams.write
    get:
      description: Watch Team
      parameters:
      - description: Team name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      - description: Team name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successful Team watch
      security:
      - oidc:
        - spirits::sessions:teams.write
        - spirits::sessions:teams.read
    put:
      description: Update Team
      parameters:
      - description: Team name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      - description: Team name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Team name
          minLength: 1
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
        description: Team to update
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successful Team update
      security:
      - oidc:
        - spirits::sessions:teams.write
  /sessions/{sessionName}/teams/{teamName}/spirits:
    get:
      description: Watch Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit watch
      security:
      - oidc:
        - spirits::sessions:teams:spirits.write
        - spirits::sessions:teams:spirits.read
    post:
      description: Create a Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Spirit'
        description: Spirit to create
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit create
      security:
      - oidc:
        - spirits::sessions:teams:spirits.write
  /sessions/{sessionName}/teams/{teamName}/spirits/{spiritName}:
    delete:
      description: Watch Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: spiritName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit delete
      security:
      - oidc:
        - spirits::sessions:teams:spirits.write
    get:
      description: Watch Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: spiritName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit watch
      security:
      - oidc:
        - spirits::sessions:teams:spirits.write
        - spirits::sessions:teams:spirits.read
    put:
      description: Update Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: teamName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: spiritName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Spirit'
        description: Spirit to update
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit update
      security:
      - oidc:
        - spirits::sessions:teams:spirits.write
...
  /sessions/{sessionName}/battles/{battleName}/spirits:
    get:
      description: Watch Spirit
      parameters:
      - description: Spirit name
        explode: false
        in: path
        name: sessionName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      - description: Spirit name
        explode: false
        in: path
        name: battleName
        required: true
        schema:
          description: Spirit name
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spirit'
          description: Successful Spirit watch
      security:
      - oidc:
        - spirits::sessions:battles:spirits.write
        - spirits::sessions:battles:spirits.read
...
components:
  schemas:
    Session:
      description: An isolated collection of Battle's and Team's
      example:
        auth:
          oidc:
            issuer: issuer
        name: name
      properties:
        name:
          description: The unique name of this Session
          minLength: 1
          type: string
        auth:
          $ref: '#/components/schemas/Session_auth'
      required:
      - name
      type: object
    Team:
      description: A collection Spirit's
      example:
        name: name
      properties:
        name:
          description: The unique name of this Team
          minLength: 1
          type: string
      required:
      - name
      type: object
    Spirit:
      description: An actor in a Battle
      example:
        stats:
          armor: 0
          health: 1
          power: 0
          agility: 0
        name: name
        actions:
        - actions
        - actions
        intelligence: roundRobin
      properties:
        name:
          description: The unique name of this Spirit
          minLength: 1
          type: string
        stats:
          $ref: '#/components/schemas/Spirit_stats'
        actions:
          default:
          - attack
          description: The Action's that this Spirit can perform
          items:
            description: The name of an Action
            type: string
          type: array
        intelligence:
          default: roundRobin
          description: The AI setting for this Spirit
          type: string
      required:
      - name
      - stats
      type: object
...
  securitySchemes:
    oidc:
      openIdConnectUrl: https://oh-great-spirits.herokuapp.com
      type: openIdConnect